+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (@egodigital/appstore-connect)
 
+## 0.2.0
+
+* added filter property to [SalesReportRow](https://egodigital.github.io/appstore-connect/interfaces/_index_.salesreportrow.html) interface
+* updated the following npm modules:
+  * [got](https://npmjs.com/package/got): `^10.2.2`
+
 ## 0.1.5
 
 * bug fixes